refactor: consolidate fallback mechanism tests into test_client.py and remove redundant test_fallback.py
